week13/hhback/requirements.txt
appdirs==1.4.3
astroid==2.3.3
colorama==0.4.3
distlib==0.3.0
Django==2.2.11
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
filelock==3.0.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.15
wrapt==1.11.2
